tube-dbus: subclass TpBaseChannel
This one is complicated because the old GabbleTubeDBus handled both
TargetHandleType as Contact *and* Room. TpBaseChannel sets the
TargetHandleType per class so makes this hard. The easiest way to do
this is to have a really dummy subclass for MUCs to change the
TargetHandleType (and interfaces, to add Group).

It would be nice in future to move the actual MUC stuff into
GabbleMucTubeDbus, but that's for another day.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>