console UI: quit cleanly if connection has gone away
The console UI doesn't watch for NameOwnerChanged to notice the
connection going away from beneath it. No big deal, it's a tool for
nerds, but if you tried to quit it, it would not close cleanly: the call
to self.snoopy.teardown() (which tries to turn off the connection
dumping all stanzas over D-Bus) throws an exception, and so
Gtk.main_quit() never got called.