JingleSession: delegate capability queries to factory
The two remaining bits of code that use GabbleConnection inside
GabbleJingleSession use the caps cache and the roster. There is really
no good code for these inside Wocky just now (well, WockyRoster exists,
but it's not what Gabble uses so the point is moot).

I've resigned myself to the fact that GabbleJingleFactory will have to
stick around in some form—not least to listen for ::porter-available,
and also to dig stun server parameters out of GabbleConnection. So I
think the best path to breaking GabbleJingleSession's Gabble addiction
is to add some signals which get handled by the Gabble-specific remnants
of the JingleFactory.

Meanwhile, I noticed that the caps-checking code didn't actually work
for the bare-JID case, so fixed that up while moving the code.