JingleMediaRtp: drop use of TP_ERRORS
Depending on whether an illegal codec update comes from the network or
from the local system, we end up wanting either a WOCKY_XMPP_ERROR or a
TP_ERROR (to throw back over the wire or in a D-Bus method error,
respectively). Previously, this choice was made inside JingleMediaRtp;
this patch moves the conversion out to GabbleMediaStream. (The Call
codepaths all pass NULL for the GError **... great.)