Wrap JingleFactory in a new class, JingleMint.
The intention is to move all Gabble-specific code into JingleMint,
leaving only the pure angle-bracket distillate behind in JingleFactory,
which can then be moved to Wocky. But as of this commit, JingleMint is
basically an empty shell: it immediately creates a JingleFactory, and
proxies its ::new-session signal.

As described in the comments, “mint” is meant in the “factory which
makes coins” sense. It seemed like a better name than
JingleFactoryFactory.