bytestream-socks5: send empty offer if we can't initiate.
This looks pretty ridiculous, so needs some explanation. If you have no
network connection (suppose you're on a train, testing tubes using
Prosody on localhost), then get_local_interfaces_ips() returns NULL
(the loopback interface is explicitly skipped). So previously, after the
recipient has accepted the tube, gabble_bytestream_socks5_initiate()
would fail for the initiator. bytestream-multiple doesn't fall back if
it fails immediately; so the whole bytestream-establishment fails. The
initiator's tube channel would close, but no stanza gets sent to the
peer to tell it to stop waiting for the bytestream offer, so it's just
hanging there waiting forever.

Just making bytestream-multiple fall back immediately doesn't help,
because the peer is still expecting a SOCKS5 offer, rather than a
in-band bytestream offer, so rejects it. Unlike Jingle, SI doesn't have
a way to cancel an offer once it's been accepted but before the
transport has been negotiated. So… if we send the peer an empty offer,
they'll send back an error, and both sides know to fall back (if they
understand bytestream-multiple) or give up (if not).

There is no test case for this, because it's difficult to make
get_local_interfaces_ips() or gibber_listener_listen_tcp() fail from the
test suite. I tested it with a pair of simple D-Bus tube clients and no
network connection, and/or hacking get_local_interfaces_ips() to always
return NULL.

Although this trick is schema-compliant
<http://xmpp.org/extensions/xep-0065.html#schema>, it's not compliant
with the wording of XEP-0065. Quoth §5.3.1
<http://xmpp.org/extensions/xep-0065.html#direct-proto-initiate>:

> The <query/> element MUST contain one or more <streamhost/> elements,
> each of which MUST possess the 'host', 'jid', and 'port' attributes.

But it also says:

> If the request is malformed …, the Target MUST return an error of
> <bad-request/>.

So this should work with any (robust) implementation.

The option would be to (try to) start listening before sending the
SI offer—if it fails, don't offer to use SOCKS5 bytestreams.

https://bugs.freedesktop.org/show_bug.cgi?id=48050