im-channel: send delivery reports when requested...
...and the sender is able to see that we are online, is not blocked, and
so on. This is not a presence leak: all this allows the peer to do that
they cannot already do is determine that our internet connection has not
died without our server noticing.

Sending a receipt as soon as the message is signalled to the UI, rather
than after it's acknowledged, is absolutely compliant with XEP-0184 §4:

> Finally, this protocol does not enable the sender to know that the intended
> recipient has read the message[, ...] that an end user client has
> presented the message to a human user (if any), etc.