sidecars: fix enqueuing requests made before connecting
tp_base_connection_get_status() returns DISCONNECTED in two states: when
the connection is yet to connect (internally, it is in state NEW) and
when the connection is defunct. tp_base_connection_is_destroyed() allows
us to distinguish the two in this case. This was introduced in commit
fde8437 and was caught by gateways.py.

Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=55908