Update Wocky snapshot to validate stanza namespaces
This breaks the console plugin, which checks if an entered stanza is of
a known type, but <message xmlns=''> is not the same as <message
xmlns='jabber:client'> so Wocky now says the former has type UNKNOWN.
The plugin already had some code to fix up empty namespaces, but it's
after the type check. For a better fix, I added API to give
non-streaming WockyXmppReaders a default namespace, and used it here.

In the course of fixing this, I found that telling the console to send
this:

   <message>
     <body>
       hai
    </body>
  </message>

would send this:

   <message xmlns='jabber:client'>
     <body xmlns=''>
       hai
    </body>
  </message>

which is wrong: the empty namespace was not being fixed up recursively.
This is fixed as a side-effect of the default-namespace property, but
this patch also adds a test.

https://bugs.freedesktop.org/show_bug.cgi?id=57016