auth-manager: simplify chaining up in _start_auth_async()
Rather than using whether or not priv->channel is NULL in
_start_auth_finish() to determine how to handle the result, let's always
keep our own GSimpleAsyncResult.

In this case it doesn't matter very much, because priv->channel in
practice can't become NULL in the gap between the callback being called
and Wocky calling _start_auth_finish(). In the other cases it does
matter, and causes crashes. I'll be fixing those next.