auth-manager: propagate failure both up and across
If we're delegating up to Wocky's internal SASL machinery, it might want
to know about the handshake failing.

If a SASL channel is open, we should let it know so that it can let its
handler know that the handshake failed.

If we're using X-TELEPATHY-PASSWORD, both these conditions are true.