jingle/stun-server: remove unneeded do_connect=False
This used to be needed to ensure that the query Gabble sends right after
connecting wasn't swallowed by the expect() calls that gabbletest makes
after Connect(). But now that Sjoerd's split the dbus and stream event
queues, this is no longer a problem: the only stream event it awaits is
stream-authenticated, which is certainly before the jingleinfo query.