gabbletest: properly support domains other than @localhost
I need to be able to vary the domain to do a variety of STUN SRV tests.
Previously, whatever domain Gabble tried to connect as, the server would
just bind you back to @localhost. The only test which tried to do
something exotic was the Facebook own-message test, and that one seemed
to work anyway… oh well!