vcard-manager: fix is_significant()
gabble_vcard_manager_replace_is_significant() is meant to check whether
we're actually changing a value, but in fact it basically always claimed
we're not, because it forgot to make a note of whether it's seen the
relevant element.

By fixing this, we can remove a redundant check at the end of
gabble_vcard_manager_edit_info_apply().