test-multift: don't rely on q.expect timing out
Previously, this test checked that events didn't happen by waiting for
them to happen, asserting if they did, and catching the TimeoutError
which occurred if they didn't.

But it's more obvious what's happening to forbid the event, take all the
actions, and then sync the stream to make sure it hasn't appeared. Plus,
it shaves 6 seconds off the test.