tests: don't crash in verbose mode on unicode in stanzas
We have this hack that sets __repr__ on domish.Element to call its
toXml() method, which ends up being used by the verbose logging code
when it dumps all of an event's attributes. Unfortunately, this blows up
if a stanza contains non-ASCII characters, because repr() tries to
convert unicode to str using .encode('ascii').

This made running jingle/test-send-file in verbose mode fail, because
the file being sent by the test has a non-ASCII filename. The fix is to
make the __repr__ hack escape non-ASCII characters using the
unicode-escape codec.