roster/groups: don't rely on an implementation detail
Before telepathy-glib 0.20.3 and 0.21.1, we had this incorrect sequence
(pseudocode) for each group:

* NewChannels(the group)
* GroupsChanged([the group], added: [...], removed: [])
* NewChannels(the group)

In 0.20.3 and 0.20.1, we removed the second emission of NewChannels.
Unfortunately, that broke this test, which was specifically expecting
GroupsChanged followed by NewChannels.

Rather than reversing the assumption, I'm doing it properly, by
expecting the events in no particular order.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=67828
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>