roster: stop using group handles
Group handles will be removed in Telepathy 1.0 so we should stop using them to
as an implementation detail in roster.c

I did this by replacing TpHandleSet/TpIntset by GHashTable of (gchar *group)
used as a set. I re-implemented the logic of some tp_intset_* methods to keep
the existing logic of the code.