tests: work in terms of G_TEST_SRCDIR, G_TEST_BUILDDIR and ${twistedtestsdir}
GLib already has a convention that G_TEST_SRCDIR, G_TEST_BUILDDIR
are set to @abs_srcdir@, @abs_builddir@. Let's use it.

That makes it much simpler to use a substitution variable whose value
is .../tests/twisted rather than .../tests, so do that too. As a side
benefit, this removes a variable whose name is Gabble-specific.