tests: give signature explicitly when calling optional interfaces
Our regression tests use dbus-python, which uses
Introspectable.Introspect() when the proxy is first constructed.
When telepathy-glib passed a static list of "every possible interface"
to either dbus-glib or GDBus on object-creation, this was sufficient.

However, now that we only add interfaces when we discover
that the server can support them, those interfaces won't have been in
the initial Introspect() result. We could construct a new proxy to
force re-introspection, but it's easier (and nicer) to be explicit
about types.