all TpChannelManager implementations: remove before emitting ChannelClosed
When TpBaseConnection updates its idea of the Channels property
in response to our channel-closed signal, we don't want it to list the
just-closed channel in foreach_channel().

Similarly, make sure we survive foreach_channel() being called during
disconnection, even if we have destroyed the data structure storing
our channels.

The IM and MUC managers are (as usual) more subtle than the rest
because of their "respawning" behaviour: if we're respawning channels,
we need to emit channel-closed before new-channel, otherwise the D-Bus
signals will be in the wrong order (but if we're not, we need to emit
channel-closed last).

The FT manager didn't actually emit channel-closed at all, so fix that.